Fixed syntax errors, moved store connection closer to where results are used in DisplayPlaybook
